Implement dmp-based skills
We intend to obtain state-only trajectories and can re-use the same
discretization for recorded and newly generalized trajectories. The
robot controller can interpolate or sample if desired. This allows us to
use the forcing term directly, without the need to parameterize it with
Gaussian basis functions with the timeless phase variable.  Spatial
scaling works just as well with numerical integration and temporal
scaling works with simple re-writing the time stamps of the obtained
trajectories.

If, at some point, we need just-in-time trajectory execution with
different durations, then we can implement the full DMP approach.

Also add a test for learning trajectories.