Initial commit for the project
This is in fact an older state in which I wanted to implement robot
simulation in gazebo. I will change that to light-weight
ros_control_boilerplate.
I think I did something with a Gazebo camera here to check if it's
possible to give visual feedback for control.
This commit is for later reference if I need something similar later in
the project.