Use skill-local coordinates for macro generalization
The problem was that using the globally recorded start state
made it an attractor point during replay. The macro would always try to
visit this global start state before continuing the motion.

It makes more sense to record the macro locally, with respect to the
robot's end-effector frame when the recording starts. This allows the robot
to replay macros everywhere in its workspace, using only the globally
recorded goal state as attractor state. For this to work, the goal state
needs to be mapped to this skill-local frame prior to generate a new
trajectory.