Add a mode for hybrid macro replay
This new mode is a hybrid between the global and local macro execution.
In comparison to the global mode, it uses the current orientation of the
end-effector and is thus way more intuitive.

I tested this in simulation and it looks as expected.
We might need to actually measure the accuracy with TF, though. I
sometimes have the feeling that it's slightly off by few centimeters
when coming from very distant starts with completely different
orientation.

The mathematical transformations for this will need an additional
graphical illustration in later commits.