Add special macro mode to visualize motion in RViz
This is a nice helper to visualize what the generated macro motion will
look like in Cartesian space with RViz _Path_ element.
Providing a separate mode for this allows to safely inspect everything
before actually commanding a robot.
It's also good for debugging.

These changes required to move most of the _start playback_ code
into a separate function.