Fix hybrid macro execution
As it turns out, we forgot a crucial transformation and didn't test
enough with highly twisted recorded motion.
Spotting this was now easier with the visualization.

Note that this also fixes the offsets we experienced in 60c7822.
Everything works as expected now and we have *absolute* accuracy to the
global targets in hybrid macro mode.