Fix spatial scaling of trajectories
This is an important fix for really scaling the motion profile over the
complete course of the execution. Drastically longer or shorter motion
was distorted by the stiffness, which also changes with the goal
distance.

We now perform a perfect scale in the Euclidean goal direction by
rescaling relevant components of the forcing term.